[REFACTOR] 도메인형으로 변경
