+++ spring/demo/src/main/java/com/example/demo/domain/apply/controller/package-info.java
@@ -1,0 +1,1 @@
+package com.example.demo.domain.apply.controller;
