+++ spring/demo/src/main/java/com/example/demo/domain/event/controller/package-info.java
@@ -1,0 +1,1 @@
+package com.example.demo.domain.event.controller;
