+++ spring/demo/src/main/java/com/example/demo/domain/ticket/repository/package-info.java
@@ -1,0 +1,1 @@
+package com.example.demo.domain.ticket.repository;
