--- spring/demo/src/main/java/com/example/demo/event/service/package-info.java
@@ -1,1 +1,0 @@
-package com.example.demo.event.service;
