--- spring/demo/src/main/java/com/example/demo/ticket/controller/package-info.java
@@ -1,1 +1,0 @@
-package com.example.demo.ticket.controller;
