--- spring/demo/src/main/java/com/example/demo/user/repository/package-info.java
@@ -1,1 +1,0 @@
-package com.example.demo.user.repository;
